Implement ir unit deserialization methods
