+++ components/core/src/clp/ffi/ir_stream/ir_unit_deserialization_methods.cpp
@@ -1,0 +1,535 @@
+#include "ir_unit_deserialization_methods.hpp"
+
+#include <cstddef>
+#include <cstdint>
+#include <memory>
+#include <optional>
+#include <string>
+#include <type_traits>
+#include <utility>
+#include <vector>
+
+#include <outcome/single-header/outcome.hpp>
+
+#include "../../ErrorCode.hpp"
+#include "../../ir/EncodedTextAst.hpp"
+#include "../../ir/types.hpp"
+#include "../../ReaderInterface.hpp"
+#include "../../time_types.hpp"
+#include "../../type_utils.hpp"
+#include "../KeyValuePairLogEvent.hpp"
+#include "../SchemaTree.hpp"
+#include "../SchemaTreeNode.hpp"
+#include "../Value.hpp"
+#include "decoding_methods.hpp"
+#include "protocol_constants.hpp"
+#include "utils.hpp"
+
+namespace clp::ffi::ir_stream {
+namespace {
+/**
+ * A collection of schema tree leaf node IDs. It represents the schema of a `KeyValuePairLogEvent`.
+ */
+using Schema = std::vector<SchemaTreeNode::id_t>;
+
+/**
+ * @param tag
+ * @return The corresponding schema tree node type on success.
+ * @return std::nullopt if the tag doesn't match to any defined schema tree node type.
+ */
+[[nodiscard]] auto schema_tree_node_tag_to_type(encoded_tag_t tag
+) -> std::optional<SchemaTreeNode::Type>;
+
+/**
+ * Deserializes the parent ID of a schema tree node.
+ * @param reader
+ * @param parent_id Returns the deserialized result.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return IRErrorCode::IRErrorCode_Incomplete_IR if the stream is truncated.
+ * @return IRErrorCode::IRErrorCode_Corrupted_IR if the next packet in the stream isn't a parent ID.
+ * @return Same as `deserialize_tag` on any other failure.
+ */
+[[nodiscard]] auto deserialize_schema_tree_node_parent_id(
+        ReaderInterface& reader,
+        SchemaTreeNode::id_t& parent_id
+) -> IRErrorCode;
+
+/**
+ * Deserializes the key name of a schema tree node.
+ * @param reader
+ * @param key_name Returns the deserialized key name.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return Same as `deserialize_tag` or `deserialize_string` on failure.
+ */
+[[nodiscard]] auto deserialize_schema_tree_node_key_name(
+        ReaderInterface& reader,
+        std::string& key_name
+) -> IRErrorCode;
+
+/**
+ * Deserializes an integer value packet.
+ * @param reader
+ * @param tag
+ * @param val Returns the deserialized value.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return IRErrorCode::IRErrorCode_Incomplete_IR if the stream is truncated.
+ * @return IRErrorCode::IRErrorCode_Corrupted_IR if the given tag doesn't correspond to an integer
+ * packet.
+ */
+[[nodiscard]] auto
+deserialize_int_val(ReaderInterface& reader, encoded_tag_t tag, value_int_t& val) -> IRErrorCode;
+
+/**
+ * Deserializes a string packet.
+ * @param reader
+ * @param tag
+ * @param deserialized_str Returns the deserialized string.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return IRErrorCode::IRErrorCode_Incomplete_IR if the stream is truncated.
+ * @return IRErrorCode::IRErrorCode_Corrupted_IR if the given tag doesn't correspond to a string
+ * packet.
+ */
+[[nodiscard]] auto deserialize_string(
+        ReaderInterface& reader,
+        encoded_tag_t tag,
+        std::string& deserialized_str
+) -> IRErrorCode;
+
+/**
+ * Deserializes the IDs of all keys in a log event.
+ * @param reader
+ * @param tag Takes the current tag as input and returns the last tag read.
+ * @param schema Returns the deserialized schema.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return IRErrorCode::IRErrorCode_Incomplete_IR if the stream is truncated.
+ * @return Same as `deserialize_tag` on any other failure.
+ */
+[[nodiscard]] auto
+deserialize_schema(ReaderInterface& reader, encoded_tag_t& tag, Schema& schema) -> IRErrorCode;
+
+/**
+ * Deserializes the next value and pushes the result into `node_id_value_pairs`.
+ * @param reader
+ * @param tag
+ * @param node_id The node ID that corresponds to the value.
+ * @param node_id_value_pairs Returns the ID-value pair constructed from the deserialized value.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return IRErrorCode::IRErrorCode_Incomplete_IR if the stream is truncated.
+ * @return IRErrorCode::IRErrorCode_Corrupted_IR if the tag doesn't correspond to any known value
+ * type.
+ * @return Same as `deserialize_encoded_text_ast_and_insert_to_node_id_value_pairs` on any other
+ * failure.
+ */
+[[nodiscard]] auto deserialize_value_and_insert_to_node_id_value_pairs(
+        ReaderInterface& reader,
+        encoded_tag_t tag,
+        SchemaTreeNode::id_t node_id,
+        KeyValuePairLogEvent::NodeIdValuePairs& node_id_value_pairs
+) -> IRErrorCode;
+
+/**
+ * Deserializes an encoded text AST and pushes the result into node_id_value_pairs.
+ * @tparam encoded_variable_t
+ * @param reader
+ * @param node_id The node ID that corresponds to the value.
+ * @param node_id_value_pairs Returns the ID-value pair constructed by the deserialized encoded text
+ * AST.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return Same as `deserialize_tag` or `deserialize_encoded_text_ast` on failure.
+ */
+template <typename encoded_variable_t>
+requires(std::is_same_v<ir::four_byte_encoded_variable_t, encoded_variable_t>
+         || std::is_same_v<ir::eight_byte_encoded_variable_t, encoded_variable_t>)
+[[nodiscard]] auto deserialize_encoded_text_ast_and_insert_to_node_id_value_pairs(
+        ReaderInterface& reader,
+        SchemaTreeNode::id_t node_id,
+        KeyValuePairLogEvent::NodeIdValuePairs& node_id_value_pairs
+) -> IRErrorCode;
+
+/**
+ * Deserializes values and constructs ID-value pairs according to the given schema. The number of
+ * values to deserialize is indicated by the size of the given schema.
+ * @param reader
+ * @param tag
+ * @param schema The log event's schema.
+ * @param node_id_value_pairs Returns the constructed ID-value pairs.
+ * @return IRErrorCode::IRErrorCode_Success on success.
+ * @return IRErrorCode::IRErrorCode_Corrupted_IR if a key is duplicated in the deserialized log
+ * event.
+ * @return Same as `deserialize_tag` or `deserialize_value_and_insert_to_node_id_value_pairs` on any
+ * other failure.
+ */
+[[nodiscard]] auto deserialize_value_and_construct_node_id_value_pairs(
+        ReaderInterface& reader,
+        encoded_tag_t tag,
+        Schema const& schema,
+        KeyValuePairLogEvent::NodeIdValuePairs& node_id_value_pairs
+) -> IRErrorCode;
+
+auto schema_tree_node_tag_to_type(encoded_tag_t tag) -> std::optional<SchemaTreeNode::Type> {
+    switch (tag) {
+        case cProtocol::Payload::SchemaTreeNodeInt:
+            return SchemaTreeNode::Type::Int;
+        case cProtocol::Payload::SchemaTreeNodeFloat:
+            return SchemaTreeNode::Type::Float;
+        case cProtocol::Payload::SchemaTreeNodeBool:
+            return SchemaTreeNode::Type::Bool;
+        case cProtocol::Payload::SchemaTreeNodeStr:
+            return SchemaTreeNode::Type::Str;
+        case cProtocol::Payload::SchemaTreeNodeUnstructuredArray:
+            return SchemaTreeNode::Type::UnstructuredArray;
+        case cProtocol::Payload::SchemaTreeNodeObj:
+            return SchemaTreeNode::Type::Obj;
+        default:
+            return std::nullopt;
+    }
+}
+
+auto deserialize_schema_tree_node_parent_id(
+        ReaderInterface& reader,
+        SchemaTreeNode::id_t& parent_id
+) -> IRErrorCode {
+    encoded_tag_t tag{};
+    if (auto const err{deserialize_tag(reader, tag)}; IRErrorCode::IRErrorCode_Success != err) {
+        return err;
+    }
+    if (cProtocol::Payload::SchemaTreeNodeParentIdUByte == tag) {
+        uint8_t deserialized_id{};
+        if (false == deserialize_int(reader, deserialized_id)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        parent_id = static_cast<SchemaTreeNode::id_t>(deserialized_id);
+    } else if (cProtocol::Payload::SchemaTreeNodeParentIdUShort == tag) {
+        uint16_t deserialized_id{};
+        if (false == deserialize_int(reader, deserialized_id)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        parent_id = static_cast<SchemaTreeNode::id_t>(deserialized_id);
+    } else {
+        return IRErrorCode::IRErrorCode_Corrupted_IR;
+    }
+    return IRErrorCode_Success;
+}
+
+auto deserialize_schema_tree_node_key_name(ReaderInterface& reader, std::string& key_name)
+        -> IRErrorCode {
+    encoded_tag_t str_packet_tag{};
+    if (auto const err{deserialize_tag(reader, str_packet_tag)};
+        IRErrorCode::IRErrorCode_Success != err)
+    {
+        return err;
+    }
+    if (auto const err{deserialize_string(reader, str_packet_tag, key_name)};
+        IRErrorCode::IRErrorCode_Success != err)
+    {
+        return err;
+    }
+    return IRErrorCode::IRErrorCode_Success;
+}
+
+auto deserialize_int_val(ReaderInterface& reader, encoded_tag_t tag, value_int_t& val)
+        -> IRErrorCode {
+    if (cProtocol::Payload::ValueInt8 == tag) {
+        int8_t deserialized_val{};
+        if (false == deserialize_int(reader, deserialized_val)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        // NOLINTNEXTLINE(bugprone-signed-char-misuse,cert-str34-c)
+        val = deserialized_val;
+    } else if (cProtocol::Payload::ValueInt16 == tag) {
+        int16_t deserialized_val{};
+        if (false == deserialize_int(reader, deserialized_val)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        val = deserialized_val;
+    } else if (cProtocol::Payload::ValueInt32 == tag) {
+        int32_t deserialized_val{};
+        if (false == deserialize_int(reader, deserialized_val)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        val = deserialized_val;
+    } else if (cProtocol::Payload::ValueInt64 == tag) {
+        int64_t deserialized_val{};
+        if (false == deserialize_int(reader, deserialized_val)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        val = deserialized_val;
+    } else {
+        return IRErrorCode::IRErrorCode_Corrupted_IR;
+    }
+    return IRErrorCode::IRErrorCode_Success;
+}
+
+auto deserialize_string(ReaderInterface& reader, encoded_tag_t tag, std::string& deserialized_str)
+        -> IRErrorCode {
+    size_t str_length{};
+    if (cProtocol::Payload::StrLenUByte == tag) {
+        uint8_t length{};
+        if (false == deserialize_int(reader, length)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        str_length = static_cast<size_t>(length);
+    } else if (cProtocol::Payload::StrLenUShort == tag) {
+        uint16_t length{};
+        if (false == deserialize_int(reader, length)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        str_length = static_cast<size_t>(length);
+    } else if (cProtocol::Payload::StrLenUInt == tag) {
+        uint32_t length{};
+        if (false == deserialize_int(reader, length)) {
+            return IRErrorCode::IRErrorCode_Incomplete_IR;
+        }
+        str_length = static_cast<size_t>(length);
+    } else {
+        return IRErrorCode::IRErrorCode_Corrupted_IR;
+    }
+    if (clp::ErrorCode_Success != reader.try_read_string(str_length, deserialized_str)) {
+        return IRErrorCode::IRErrorCode_Incomplete_IR;
+    }
+    return IRErrorCode::IRErrorCode_Success;
+}
+
+auto deserialize_schema(ReaderInterface& reader, encoded_tag_t& tag, Schema& schema)
+        -> IRErrorCode {
+    schema.clear();
+    while (true) {
+        if (cProtocol::Payload::KeyIdUByte == tag) {
+            uint8_t id{};
+            if (false == deserialize_int(reader, id)) {
+                return IRErrorCode::IRErrorCode_Incomplete_IR;
+            }
+            schema.push_back(static_cast<SchemaTreeNode::id_t>(id));
+        } else if (cProtocol::Payload::KeyIdUShort == tag) {
+            uint16_t id{};
+            if (false == deserialize_int(reader, id)) {
+                return IRErrorCode::IRErrorCode_Incomplete_IR;
+            }
+            schema.push_back(static_cast<SchemaTreeNode::id_t>(id));
+        } else {
+            break;
+        }
+
+        if (auto const err{deserialize_tag(reader, tag)}; IRErrorCode::IRErrorCode_Success != err) {
+            return err;
+        }
+    }
+
+    return IRErrorCode::IRErrorCode_Success;
+}
+
+auto deserialize_value_and_insert_to_node_id_value_pairs(
+        ReaderInterface& reader,
+        encoded_tag_t tag,
+        SchemaTreeNode::id_t node_id,
+        KeyValuePairLogEvent::NodeIdValuePairs& node_id_value_pairs
+) -> IRErrorCode {
+    switch (tag) {
+        case cProtocol::Payload::ValueInt8:
+        case cProtocol::Payload::ValueInt16:
+        case cProtocol::Payload::ValueInt32:
+        case cProtocol::Payload::ValueInt64: {
+            value_int_t value_int{};
+            if (auto const err{deserialize_int_val(reader, tag, value_int)};
+                IRErrorCode::IRErrorCode_Success != err)
+            {
+                return err;
+            }
+            node_id_value_pairs.emplace(node_id, Value{value_int});
+            break;
+        }
+        case cProtocol::Payload::ValueFloat: {
+            uint64_t val{};
+            if (false == deserialize_int(reader, val)) {
+                return IRErrorCode::IRErrorCode_Incomplete_IR;
+            }
+            node_id_value_pairs.emplace(node_id, Value{bit_cast<value_float_t>(val)});
+            break;
+        }
+        case cProtocol::Payload::ValueTrue:
+            node_id_value_pairs.emplace(node_id, Value{true});
+            break;
+        case cProtocol::Payload::ValueFalse:
+            node_id_value_pairs.emplace(node_id, Value{false});
+            break;
+        case cProtocol::Payload::StrLenUByte:
+        case cProtocol::Payload::StrLenUShort:
+        case cProtocol::Payload::StrLenUInt: {
+            std::string value_str;
+            if (auto const err{deserialize_string(reader, tag, value_str)};
+                IRErrorCode::IRErrorCode_Success != err)
+            {
+                return err;
+            }
+            node_id_value_pairs.emplace(node_id, Value{std::move(value_str)});
+            break;
+        }
+        case cProtocol::Payload::ValueEightByteEncodingClpStr:
+            if (auto const err{deserialize_encoded_text_ast_and_insert_to_node_id_value_pairs<
+                        ir::eight_byte_encoded_variable_t>(reader, node_id, node_id_value_pairs)};
+                IRErrorCode::IRErrorCode_Success != err)
+            {
+                return err;
+            }
+            break;
+        case cProtocol::Payload::ValueFourByteEncodingClpStr:
+            if (auto const err{deserialize_encoded_text_ast_and_insert_to_node_id_value_pairs<
+                        ir::four_byte_encoded_variable_t>(reader, node_id, node_id_value_pairs)};
+                IRErrorCode::IRErrorCode_Success != err)
+            {
+                return err;
+            }
+            break;
+        case cProtocol::Payload::ValueNull:
+            node_id_value_pairs.emplace(node_id, Value{});
+            break;
+        case cProtocol::Payload::ValueEmpty:
+            node_id_value_pairs.emplace(node_id, std::nullopt);
+            break;
+        default:
+            return IRErrorCode::IRErrorCode_Corrupted_IR;
+    }
+    return IRErrorCode::IRErrorCode_Success;
+}
+
+template <typename encoded_variable_t>
+requires(std::is_same_v<ir::four_byte_encoded_variable_t, encoded_variable_t>
+         || std::is_same_v<ir::eight_byte_encoded_variable_t, encoded_variable_t>)
+[[nodiscard]] auto deserialize_encoded_text_ast_and_insert_to_node_id_value_pairs(
+        ReaderInterface& reader,
+        SchemaTreeNode::id_t node_id,
+        KeyValuePairLogEvent::NodeIdValuePairs& node_id_value_pairs
+) -> IRErrorCode {
+    encoded_tag_t tag{};
+    if (auto const err{deserialize_tag(reader, tag)}; IRErrorCode::IRErrorCode_Success != err) {
+        return err;
+    }
+
+    std::string logtype;
+    std::vector<encoded_variable_t> encoded_vars;
+    std::vector<std::string> dict_vars;
+    if (auto const err{deserialize_encoded_text_ast(reader, tag, logtype, encoded_vars, dict_vars)};
+        IRErrorCode::IRErrorCode_Success != err)
+    {
+        return err;
+    }
+
+    node_id_value_pairs.emplace(
+            node_id,
+            Value{ir::EncodedTextAst<encoded_variable_t>{logtype, dict_vars, encoded_vars}}
+    );
+    return IRErrorCode::IRErrorCode_Success;
+}
+
+auto deserialize_value_and_construct_node_id_value_pairs(
+        ReaderInterface& reader,
+        encoded_tag_t tag,
+        Schema const& schema,
+        KeyValuePairLogEvent::NodeIdValuePairs& node_id_value_pairs
+) -> IRErrorCode {
+    node_id_value_pairs.clear();
+    node_id_value_pairs.reserve(schema.size());
+    for (auto const node_id : schema) {
+        if (node_id_value_pairs.contains(node_id)) {
+            // The key should be unique in a schema
+            return IRErrorCode_Corrupted_IR;
+        }
+
+        if (auto const err{deserialize_value_and_insert_to_node_id_value_pairs(
+                    reader,
+                    tag,
+                    node_id,
+                    node_id_value_pairs
+            )};
+            IRErrorCode::IRErrorCode_Success != err)
+        {
+            return err;
+        }
+
+        if (schema.size() != node_id_value_pairs.size()) {
+            if (auto const err{deserialize_tag(reader, tag)};
+                IRErrorCode::IRErrorCode_Success != err)
+            {
+                return err;
+            }
+        }
+    }
+    return IRErrorCode::IRErrorCode_Success;
+}
+}  // namespace
+
+auto deserialize_ir_unit_schema_tree_node_insertion(
+        ReaderInterface& reader,
+        encoded_tag_t tag,
+        std::string& key_name
+) -> OUTCOME_V2_NAMESPACE::std_result<SchemaTree::NodeLocator> {
+    auto const type{schema_tree_node_tag_to_type(tag)};
+    if (false == type.has_value()) {
+        return ir_error_code_to_errc(IRErrorCode::IRErrorCode_Corrupted_IR);
+    }
+
+    SchemaTreeNode::id_t parent_id{};
+    if (auto const err{deserialize_schema_tree_node_parent_id(reader, parent_id)};
+        IRErrorCode_Success != err)
+    {
+        return ir_error_code_to_errc(err);
+    }
+
+    if (auto const err{deserialize_schema_tree_node_key_name(reader, key_name)};
+        IRErrorCode::IRErrorCode_Success != err)
+    {
+        return ir_error_code_to_errc(err);
+    }
+
+    return SchemaTree::NodeLocator{parent_id, key_name, type.value()};
+}
+
+auto deserialize_ir_unit_utc_offset_change(ReaderInterface& reader
+) -> OUTCOME_V2_NAMESPACE::std_result<UtcOffset> {
+    UtcOffset utc_offset{0};
+    if (auto const err{deserialize_utc_offset_change(reader, utc_offset)};
+        IRErrorCode::IRErrorCode_Success != err)
+    {
+        return ir_error_code_to_errc(err);
+    }
+    return utc_offset;
+}
+
+auto deserialize_ir_unit_kv_pair_log_event(
+        ReaderInterface& reader,
+        encoded_tag_t tag,
+        std::shared_ptr<SchemaTree> schema_tree,
+        UtcOffset utc_offset
+) -> OUTCOME_V2_NAMESPACE::std_result<KeyValuePairLogEvent> {
+    Schema schema;
+    if (auto const err{deserialize_schema(reader, tag, schema)};
+        IRErrorCode::IRErrorCode_Success != err)
+    {
+        return ir_error_code_to_errc(err);
+    }
+
+    KeyValuePairLogEvent::NodeIdValuePairs node_id_value_pairs;
+    if (false == schema.empty()) {
+        if (auto const err{deserialize_value_and_construct_node_id_value_pairs(
+                    reader,
+                    tag,
+                    schema,
+                    node_id_value_pairs
+            )};
+            IRErrorCode::IRErrorCode_Success != err)
+        {
+            return ir_error_code_to_errc(err);
+        }
+    } else {
+        if (cProtocol::Payload::ValueEmpty != tag) {
+            return ir_error_code_to_errc(IRErrorCode::IRErrorCode_Corrupted_IR);
+        }
+    }
+
+    return KeyValuePairLogEvent::create(
+            std::move(schema_tree),
+            std::move(node_id_value_pairs),
+            utc_offset
+    );
+}
+}  // namespace clp::ffi::ir_stream
